+++ messages/ru-RU/app/modules/guard.php
@@ -29,6 +29,14 @@ return [
     'Unbanned' => "Разблокирован",
     'Released' => "Освобождён",
 
+    'Block by IP or network' => "Заблокировать по IP или сети",
+    '{count} addresses added successfully!' => "{count} адрес(ов) успешно добавлены!",
+    '{count} addresses were added successfully, but some errors occurred: {errors}' => "{count} адрес(ов) добавлено успешно, но возникли некоторые ошибки: {errors}",
+    'An error occurred while add the addresses: {errors}' => "Ошибка при добавлении адресов: {errors}",
+
+    'Test IP/Network' => "Тестировать IP/Cеть",
+    'Test IP or network' => "Тестировать IP или сеть",
+
     'Ban' => "Заблокировать",
     'Unban' => "Разблокировать",
     'Release' => "Освободить",
@@ -36,6 +44,30 @@ return [
     'Close' => "Закрыть",
     'Save' => "Сохранить",
     'Apply' => "Применить",
+    'Go' => "Начать",
+
+    'Blocking period' => "Срок блокировки",
+    'Default' => "По-умолчанию",
+    '1 hour' => "1 час",
+    '6 hours' => "6 часов",
+    '1 day' => "1 день",
+    '1 week' => "1 неделя",
+    '2 weeks' => "2 недели",
+    '1 month' => "1 месяц",
+    '6 months' => "6 месяцев",
+    '1 year' => "1 год",
+    'Lifetime' => "Пожизненно",
+
+    'It looks like your IP matches the blocked `{ip}` and cannot be blocked.' => "Похоже, что Ваш IP совпадает с блокируемым `{ip}` и не может быть заблокирован.",
+    'It looks like your IP belongs to the blocked `{subnet}` subnet and cannot be blocked.' => "Похоже, что Ваш IP входит в блокируемую подсеть `{subnet}` и не может быть заблокирован.",
+    'It looks like your IP is in the blocking range `{start} - {end}` and cannot be blocked.' => "Похоже, что Ваш IP входит в блокируемый диапазон `{start} - {end}` и не может быть заблокирован.",
+
+    'Specify a list of IP addresses or networks (each address or network - on a new line). The following options are allowed:' => "Укажите список IP адресов или сетей (каждый адрес или сеть - с новой строки). Разрешены такие варианты записи как:",
+    'IPv4 address (for example: 172.104.89.12)' => "IPv4 адрес (например: 172.104.89.12)",
+    'network address in the CIDR (for example: 172.104.89.12/24)' => "адрес сети в виде CIDR (например: 172.104.89.12/24)",
+    'network address with mask 172.104.89.0/255.255.255.0' => "адрес сети с маской 172.104.89.0/255.255.255.0",
+    'address range like 172.104.89.0-172.104.89.255' => "диапазон адресов в виде 172.104.89.0-172.104.89.255",
+    'IPv6 address or network (2002::ac68:590c, 2002::ac68:5900/120)' => "IPv6 адрес или сеть (2002::ac68:590c, 2002::ac68:5900/120)",
 
     'All reasons' => "Все причины",
     'Manual blocking' => "Ручная блокировка",