+++ models/BannedForm.php
@@ -7,15 +7,19 @@ use Yii;
 use yii\base\Model;
 use yii\data\ActiveDataProvider;
 use wdmg\guard\models\Security;
-use yii\helpers\ArrayHelper;
-use yii\validators\IpValidator;
+use wdmg\helpers\ArrayHelper;
 
 /**
  * BannedForm represents the model `app\vendor\wdmg\guard\models\Security`.
  */
 class BannedForm extends Security
 {
-    public $ip;
+    public $ip = '';
+    public $reason = 'manual';
+    public $client_net = '';
+    public $user_agent = '';
+
+    private $_ip;
 
     /**
      * {@inheritdoc}
@@ -25,7 +29,8 @@ class BannedForm extends Security
         $rules = [
             [['ip', 'status'], 'required'],
             ['ip', 'trim'],
-            ['ip', 'each', 'rule' => ['ip', 'subnet' => null, 'ranges' => ['!system', '!private', 'any']], 'skipOnEmpty' => false]
+            ['ip', 'each', 'rule' => ['ip', 'subnet' => null, 'ranges' => ['!system', '!private', 'any'], 'ipv6' => false], 'skipOnEmpty' => false],
+            ['release_at', 'in', 'range' => array_keys($this->getReleases())],
         ];
         return ArrayHelper::merge(parent::rules(), $rules);
     }
@@ -40,6 +45,21 @@ class BannedForm extends Security
         if (is_string($this->ip)) {
             $list = explode("\r\n", $this->ip);
             $this->ip = array_filter(array_unique($list), 'trim');
+            $this->_ip = $this->ip;
+
+            $ips = [];
+            foreach ($this->ip as $ip) {
+                if (preg_match("/(\d{1,3}\.\d{1,3}\.\d{1,3}\.\d{1,3})\/(\d{1,3}\.\d{1,3}\.\d{1,3}\.\d{1,3})/", $ip, $mathes)) {
+                    $ips[] = trim($mathes[1]);
+                    $ips[] = trim($mathes[2]);
+                } elseif (preg_match("/(\d{1,3}\.\d{1,3}\.\d{1,3}\.\d{1,3})\-(\d{1,3}\.\d{1,3}\.\d{1,3}\.\d{1,3})/", $ip, $mathes)) {
+                    $ips[] = trim($mathes[1]);
+                    $ips[] = trim($mathes[2]);
+                } else {
+                    $ips[] = $ip;
+                }
+            }
+            $this->ip = $ips;
         }
 
         return parent::beforeValidate();
@@ -51,9 +71,144 @@ class BannedForm extends Security
     public function afterValidate()
     {
         if (is_array($this->ip)) {
-            $this->ip = implode("\r\n", $this->ip);
+            $this->ip = implode("\r\n", $this->_ip);
         }
         parent::afterValidate();
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function save($runValidation = true, $attributeNames = null)
+    {
+        $count = 0;
+        $data = [];
+        $errors = [];
+
+        // Prepare data for insert
+        foreach ($this->_ip as $key => $ip) {
+            if (preg_match("/(\d{1,3}\.\d{1,3}\.\d{1,3}\.\d{1,3})\/(\d{1,3}\.\d{1,3}\.\d{1,3}\.\d{1,3})/", $ip, $mathes)) { // 172.104.89.0/255.255.255.0
+                if ($cidr = IpAddressHelper::ip2BaseCidr(trim($mathes[1]), trim($mathes[2]))) {
+                    if ($range = IpAddressHelper::cidr2range($cidr, null)) {
+                        $data[] = [
+                            'client_ip' => null,
+                            'client_net' => $cidr,
+                            'range_start' => $range[0],
+                            'range_end' => $range[1]
+                        ];
+                    }
+                }
+            } elseif (preg_match("/(\d{1,3}\.\d{1,3}\.\d{1,3}\.\d{1,3})\-(\d{1,3}\.\d{1,3}\.\d{1,3}\.\d{1,3})/", $ip, $mathes)) { // 172.104.89.0-172.104.89.255
+                $cidrs = IpAddressHelper::range2cidrs(trim($mathes[1]), trim($mathes[2]));
+                foreach ($cidrs as $cidr) {
+                    $range = IpAddressHelper::cidr2range($cidr, null);
+                    $data[] = [
+                        'client_ip' => null,
+                        'client_net' => $cidr,
+                        'range_start' => $range[0],
+                        'range_end' => $range[1]
+                    ];
+                }
+            } elseif (preg_match("/(\d{1,3}\.\d{1,3}\.\d{1,3}\.\d{1,3})\/(\d{1,2})/", $ip, $mathes)) { // 172.104.89.12/24
+                $cidr = trim($mathes[0]);
+                if (IpAddressHelper::isValidCidr($cidr)) {
+                    if ($range = IpAddressHelper::cidr2range($cidr, null)) {
+                        $data[] = [
+                            'client_ip' => null,
+                            'client_net' => $cidr,
+                            'range_start' => $range[0],
+                            'range_end' => $range[1]
+                        ];
+                    }
+                }
+            } else if (IpAddressHelper::getIpVersion($ip, true) == "IPv4" && !IpAddressHelper::isLocalIp($ip)) {
+                $item = [];
+                $item['client_ip'] = $ip;
+                if ($netmask = IpAddressHelper::ipMask($ip, false)) {
+                    $cidr = IpAddressHelper::ip2cidr($ip, $netmask, 2);
+                    if (!$cidr)
+                        $cidr = IpAddressHelper::ip2cidr($ip, $netmask, 1);
+
+                    if ($cidr) {
+                        $item['client_net'] = $cidr;
+                        if ($range = IpAddressHelper::cidr2range($cidr, 1)) {
+                            $item['range_start'] = $range->start;
+                            $item['range_end'] = $range->end;
+                        }
+                    }
+                }
+                $data[] = $item;
+            } else {
+                continue;
+            }
+        }
+
+        // Unique data for insertion
+        $data = ArrayHelper::unique($data, ['client_ip', 'range_start', 'range_end']);
+
+        // Batch insert
+        foreach ($data as $item) {
+
+            $skip = false;
+            $user_ip = parent::getRemoteIp();
+            if (IpAddressHelper::isIpv4($user_ip)) {
+                if (isset($item['client_ip'])) {
+                    if ($user_ip == $item['client_ip']) {
+                        $errors[] = Yii::t('app/modules/guard', 'It looks like your IP matches the blocked `{ip}` and cannot be blocked.', [
+                            'ip' => $item['client_ip']
+                        ]);
+                        $skip = true;
+                    }
+                }
+
+                if (!$skip && isset($item['client_net'])) {
+                    if (IpAddressHelper::ipInCidr($user_ip, $item['client_net'])) {
+                        $errors[] = Yii::t('app/modules/guard', 'It looks like your IP belongs to the blocked `{subnet}` subnet and cannot be blocked.', [
+                            'subnet' => $item['client_net']
+                        ]);
+                        $skip = true;
+                    }
+                }
+
+                if (!$skip && isset($item['range_start']) && isset($item['range_end'])) {
+                    if (IpAddressHelper::ipInRange($user_ip, $item['range_start'], $item['range_end'])) {
+                        $errors[] = Yii::t('app/modules/guard', 'It looks like your IP is in the blocking range `{start} - {end}` and cannot be blocked.', [
+                            'start' => $item['range_start'],
+                            'end' => $item['range_end']
+                        ]);
+                        $skip = true;
+                    }
+                }
+            }
+
+            if ($skip) {
+                continue;
+            } else {
+                $security = new Security();
+                $security->reason = $this->reason;
+                $security->status = $this->status;
+                $security->client_ip = ($item['client_ip']) ? $item['client_ip'] : null;
+                $security->client_net = ($item['client_net']) ? $item['client_net'] : null;
+                $security->range_start = ($item['range_start']) ? $item['range_start'] : null;
+                $security->range_end = ($item['range_end']) ? $item['range_end'] : null;
+
+                if (is_string($this->release_at) && $this->release_at !== 'default')
+                    $security->release_at = $security->getReleaseDate(str_replace('_', ' ', $this->release_at));
+
+                if ($security->validate()) {
+                    if ($security->save()) {
+                        $count++;
+                    }
+                }
+            }
+
+            $errors = array_merge($security->errors, $errors);
+        }
+
+        return [
+            'errors' => $errors,
+            'count' => $count
+        ];
     }
 
     /**
@@ -62,7 +217,8 @@ class BannedForm extends Security
     public function attributeLabels()
     {
         $labels = [
-            'ip' => Yii::t('app/modules/guard', 'IP or/and Net')
+            'ip' => Yii::t('app/modules/guard', 'IP or/and Net'),
+            'release_at' => Yii::t('app/modules/guard', 'Blocking period')
         ];
         return ArrayHelper::merge(parent::attributeLabels(), $labels);
     }
@@ -78,6 +234,27 @@ class BannedForm extends Security
             self::GUARD_STATUS_IS_BANNED => Yii::t('app/modules/guard', 'Ban'),
             self::GUARD_STATUS_IS_UNBANNED => Yii::t('app/modules/guard', 'Unban'),
             self::GUARD_STATUS_IS_RELEASED => Yii::t('app/modules/guard', 'Release')
+        ];
+    }
+
+    /**
+     * Returns an array of blocking list statuses.
+     *
+     * @return array
+     */
+    public function getReleases()
+    {
+        return [
+            'default' => Yii::t('app/modules/guard', 'Default'),
+            '1_hour' => Yii::t('app/modules/guard', '1 hour'),
+            '6_hours' => Yii::t('app/modules/guard', '6 hours'),
+            '1_day' => Yii::t('app/modules/guard', '1 day'),
+            '1_week' => Yii::t('app/modules/guard', '1 week'),
+            '2_weeks' => Yii::t('app/modules/guard', '2 weeks'),
+            '1_month' => Yii::t('app/modules/guard', '1 month'),
+            '6_months' => Yii::t('app/modules/guard', '6 months'),
+            '1_year' => Yii::t('app/modules/guard', '1 year'),
+            'lifetime' => Yii::t('app/modules/guard', 'Lifetime')
         ];
     }
 }